move main page and user guide code into pages
