+++ frontend/constants.py
@@ -1,3 +1,8 @@
+import os
+
+# API Configuration
+API_BASE_URL = os.getenv('API_BASE_URL', 'http://django-backend:8000')
+
 # ============================================================================
 # MODEL OPTIONS
 # ============================================================================
